+++ docs/changelog.md
@@ -1,14 +1,15 @@
 
 # v7.0.0.3 Beta (2023-??-??)
 
+- New menu item `Settings/Setup/Add mpv.net to Path environment variable' added.
+- New menu item `Settings/Edit mpv.conf` added for opening mpv.conf with a text editor. Default binding `c`.
+- New menu item `Settings/Edit input.conf` added for opening input.conf with a text editor. Default binding `k`.
 - mpv.net can no longer be downloaded from the Microsoft store due
   to a general very poor experience with the package creation and submission.
   This means winget download support is unavailable until a new winget solution is implemented.
-- New menu item added at Settings/Setup to add mpv.net to the path environment variable.
 - Improved conf file reader/writer.
 - Conf editor support added for the mpv options:
   `reset-on-next-file`, `input-ipc-server`, `background`, `title`
-
 # v7.0.0.2 Beta (2023-12-13)
 
 - Besides a portable download there is now again a setup installer.